Add test cases
These test cases show that there's a problem with the precedence of
unary minus, in that it binds too weakly (so that for instance
`-(x) - y` is interpreted as `-(x - y)`, which is not in agreement with
the official CodeQL parser). Likewise `-a * b` is interpreted as
`-(a * b)`.

Interestingly, if `x` is a primitive type (so that the above snippet is
interpreted as a prefix cast), the parse tree is correct.